neuroscan/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
tensorflow==2.16.1
opencv-python==4.9.0.80
numpy==1.26.4
weasyprint==61.2
Jinja2==3.1.3
scipy==1.13.1
openai
python-dotenv
